Creating HOC with-screen-switch and putting logic of switching of screens from app to this HOC
